Added profile dropdown
Profile Dropdown Options: View Profile/Settings/LogOut + modify homepage while logged in
