+++ src/Navbar.js
@@ -1,9 +1,12 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import homieLogo from "./assets/logo-homie.png";
 import React, { useState } from "react";
-import { Button } from "react-bootstrap";
+import { Button, Dropdown } from "react-bootstrap";
 import { Link, useMatch } from "react-router-dom";
 import AuthModal from "./components/auth-component";
+import Cookies from "js-cookie";
+import { getAuth, signOut } from "firebase/auth";
+import { useNavigate } from "react-router-dom";
 
 export default function Navbar() {
   const [showAuthModal, setShowAuthModal] = useState(false);
@@ -16,6 +19,17 @@ export default function Navbar() {
     setShowAuthModal(false);
   };
 
+  const handleLogout = async () => {
+    try {
+      const auth = getAuth();
+      Cookies.remove("idToken");
+      await signOut(auth);
+      navigate("/homie-startup");
+    } catch (error) {
+      console.error("Error signing out:", error);
+    }
+  };
+
   const isAddImobRoute = useMatch("/addImob");
   const isAboutRoute = useMatch("/about");
   const isHomeRoute = useMatch("/homie-startup");
@@ -23,6 +37,7 @@ export default function Navbar() {
   const isSettings = useMatch("/settings");
   const isRentalDetails = useMatch("/rental-details");
   const isPostsRoute = useMatch("/posts");
+  const navigate = useNavigate();
 
   return (
     <div className="navbarStyle">
@@ -64,34 +79,69 @@ export default function Navbar() {
                     <Link to="/about" className="nav-item no-underline">
                       Despre noi
                     </Link>
-                    <li className="nav-item no-underline">
-                      <Button variant="link" onClick={handleOpenAuthModal}>
+                    {Cookies.get("idToken") === undefined ? (
+                      <li className="nav-item no-underline" onClick={handleOpenAuthModal}>
                         Conectează-te
-                      </Button>
-                    </li>
+                      </li>
+                    ) : (
+                      <Dropdown>
+                        <Dropdown.Toggle
+                          variant="link"
+                          id="dropdown-basic"
+                          style={{ display: 'flex', alignItems: 'center', marginTop: '-6px', marginLeft: '86px' }}
+                        >
+                          Profil
+                        </Dropdown.Toggle>
+                        <Dropdown.Menu>
+                          <Dropdown.Item as={Link} to="/profile">
+                            Vezi profil
+                          </Dropdown.Item>
+                          <Dropdown.Item as={Link} to="/settings">
+                            Setări
+                          </Dropdown.Item>
+                          <Dropdown.Item as={Link} onClick={handleLogout}>
+                            Deconectează-te
+                          </Dropdown.Item>
+                        </Dropdown.Menu>
+                      </Dropdown>
+                    )}
                   </>
                 ) : null}
 
                 {isAddImobRoute ||
-                isAboutRoute ||
-                isProfileRoute ||
-                isSettings || 
-                isPostsRoute ||
-                isRentalDetails ? (
+                  isAboutRoute ||
+                  isProfileRoute ||
+                  isSettings ||
+                  isPostsRoute ||
+                  isRentalDetails ? (
                   // Render specific content for certain pages
                   <>
-                    <Link to="settings" className="nav-item no-underline">
-                      Setari
-                    </Link>
                     <Link to="homie-startup" className="nav-item no-underline">
                       Exploreaza imobile
                     </Link>
                     <Link to="/about" className="nav-item no-underline">
                       Despre noi
                     </Link>
-                    <Link to="/profile" className="nav-item no-underline">
-                      Profil
-                    </Link>
+                    <Dropdown>
+                      <Dropdown.Toggle
+                        variant="link"
+                        id="dropdown-basic"
+                        style={{ display: 'flex', alignItems: 'center', marginTop: '-6px', marginLeft: '86px' }}
+                      >
+                        Profil
+                      </Dropdown.Toggle>
+                      <Dropdown.Menu>
+                        <Dropdown.Item as={Link} to="/profile">
+                          Vezi profil
+                        </Dropdown.Item>
+                        <Dropdown.Item as={Link} to="/settings">
+                          Setări
+                        </Dropdown.Item>
+                        <Dropdown.Item as={Link} onClick={handleLogout}>
+                          Deconectează-te
+                        </Dropdown.Item>
+                      </Dropdown.Menu>
+                    </Dropdown>
                   </>
                 ) : null}
               </ul>
