+++ src/components/profile-page.js
@@ -147,17 +147,6 @@ export const ProfilePage = () => {
     }
   };
 
-  const handleLogout = async () => {
-    try {
-      const auth = getAuth();
-      Cookies.remove("idToken");
-      await signOut(auth);
-      navigate("/homie-startup");
-    } catch (error) {
-      console.error("Error signing out:", error);
-    }
-  };
-
   return (
     <div className="tenants-profile">
       <ReviewModal
@@ -364,10 +353,6 @@ export const ProfilePage = () => {
               <div className="text-wrapper-12" >Adaugă o recenzie</div>
               <div className="rectangle-2" />
             </div>
-            </Button>
-            <Button variant="link" onClick={handleLogout}>
-              Deconectare
-            </Button>
           </div>
         </div>
       </div>
